requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
 dj-database-url==2.2.0
 django-environ==0.11.2
 djangorestframework==3.15.2
 djangorestframework-simplejwt==5.3.1
 drf-spectacular==0.27.2
 drf-yasg==1.21.7
 django-redis==5.4.0
 redis==5.0.7
 gunicorn==22.0.0
 whitenoise==6.7.0
